changed file and folder structure
