+++ src/pages/UserForm/UserForm.tsx
@@ -1,7 +1,7 @@
 import React, { Suspense } from 'react';
-import useFormData from '../hooks/useFormData';
+import useFormData from '../../hooks/useFormData';
 
-const FormContainer = React.lazy(() => import('../components/FormContainer'));
+const FormContainer = React.lazy(() => import('../../components/FormContainer/FormContainer'));
 
 const UserForm: React.FC = () => {
   const { formData, handleChange, resetForm } = useFormData();
